requirements.txt
django
djangorestframework
pytest-django
Pillow
